Migración de IDs de int a long en toda la solución
Se reemplazan todos los identificadores de tipo int por long en entidades, DTOs, comandos, queries, viewmodels y servicios de cifrado. Se actualizan controladores, lógica de acceso a datos, colecciones y tests para soportar el nuevo tipo. Este cambio amplía la capacidad de la aplicación para manejar un mayor rango de entidades.

+++ src/SFManagement.Application/Commands/ArchiveTaskCommand.cs
@@ -1,3 +1,3 @@
 namespace SFManagement.Application.Commands;
 
-public record ArchiveTaskCommand(int TaskId);
+public record ArchiveTaskCommand(long TaskId);
