+++ src/SFManagement.Application/Commands/CreateProjectCommand.cs
@@ -2,5 +2,5 @@ namespace SFManagement.Application.Commands;
 
 public record CreateProjectCommand(string Name, string? DescriptionMd)
 {
-    public int CreatedId { get; set; }
+    public long CreatedId { get; set; }
 }
