+++ src/SFManagement.Application/Queries/GetWorkersByProjectQuery.cs
@@ -2,7 +2,7 @@ using SFManagement.Application.DTOs;
 
 namespace SFManagement.Application.Queries;
 
-public record GetWorkersByProjectQuery(int ProjectId);
+public record GetWorkersByProjectQuery(long ProjectId);
 
 public interface IGetWorkersByProjectQueryHandler
 {
